// apps/api/src/modules/invoices/dto/create-invoice.dto.d.ts
export declare class CreateInvoiceDto {
    agreementId: string;
    contactId: string;
    subtotal: number;
    taxAmount?: number;
    dueDate?: string;
}
export declare class UpdateInvoiceDto {
    subtotal?: number;
    taxAmount?: number;
    dueDate?: string;
}
//# sourceMappingURL=create-invoice.dto.d.ts.map